% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelManager.R
\name{hdp.model.export}
\alias{hdp.model.export}
\title{Export HDP Models from R to JSON}
\usage{
hdp.model.export(hdpModel, fileName)
}
\arguments{
\item{hdpModel}{An object representing your heirarchical model}

\item{file_name}{Name of the file to save. Will append .json to the end of it.}
}
\description{
This function allows you to export an HDP model as a JSON file. Please
check out th documentation on github.
}
\examples{
hdp.model.export(hdpModel, "hdpModel.json")
}
\keyword{hdm}
\keyword{hdp,}
\keyword{json,}
