% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelManager.R
\name{hdp.model.import}
\alias{hdp.model.import}
\title{Import HDP Models into R from JSON}
\usage{
hdp.model.import(jsonFile)
}
\arguments{
\item{json_file}{The file you're importing.}
}
\description{
This function allows you to import an HDP model defined in JSON. Please
check out th documentation on github.
}
\examples{
hdp.model.import("hdpModel.json")
}
\keyword{hdm}
\keyword{hdp,}
\keyword{json,}
